Some setup for rendering design tokens.
