+++ packages/docs/component-pages.njk
@@ -6,7 +6,7 @@ pagination:
     size: 1
 permalink: "components/{{ component.name | slug }}/"
 eleventyNavigation:
-  key: Chat
+  key: TODO
   parent: Components
 templateEngineOverride: njk,md
 ---
