--- packages/docs/css.md
@@ -1,10 +1,0 @@
----
-layout: layout.njk
-eleventyNavigation:
-  key: CSS
-  order: 1
----
-
-# CSS
-
-Documentation about the CSS package is coming soon.
