+++ packages/docs/css/_base.scss
@@ -42,6 +42,8 @@ h1 {
   @media screen and (min-width: 820px) {
     font-size: var(--text-xxx-large);
   }
+
+  margin-bottom: var(--space-large);
 }
 
 h2 {
