+++ packages/docs/design-tokens.md
@@ -1,0 +1,15 @@
+---
+layout: layout.njk
+eleventyNavigation:
+  key: Design tokens
+  order: 1
+templateEngineOverride: njk,md,html
+---
+
+# Design tokens
+
+Design tokens are a tech-agnostic way to store low-level values and then use them to create the styles for your website. We use tokens instead of hard coded values to ensure a scalable and consistent system.
+
+**Warning:** This page is in development.
+
+{% include 'design-tokens.njk' %}
