+++ packages/docs/package.json
@@ -4,9 +4,9 @@
   "description": "Documentation website for the U-M Library Design System.",
   "main": "index.js",
   "scripts": {
-    "prebuild": "npm run copy:components && npm run components-data",
+    "prebuild": "npm run copy:components && npm run process-data",
     "build": "npx @11ty/eleventy",
-    "components-data": "node scripts/components-data.js",
+    "process-data": "node scripts/process-data.js",
     "prestart": "npm run prebuild",
     "start": "npx @11ty/eleventy --serve",
     "copy:components": "npx mkdirp static/components && ncp node_modules/@umich-lib/components/dist static/components"
@@ -16,6 +16,7 @@
     "@11ty/eleventy-navigation": "^0.1.5",
     "@11ty/eleventy-plugin-syntaxhighlight": "^3.0.1",
     "@umich-lib/components": "^1.0.0",
+    "@umich-lib/css": "^1.0.0",
     "eleventy-plugin-nesting-toc": "^1.2.0",
     "eleventy-plugin-sass": "^1.0.0",
     "markdown-it-anchor": "^5.3.0",
